Relaxed test for flame length vs depth ignited to allow flames where the two values are equal within distance tolerance. Added new test rig to run a table of parameters read from CSV file.
